+++ client/src/components/SearchForm/index.js
@@ -1,27 +1,27 @@
-import React from "react";
-import "./style.css";
+// import React from "react";
+// import "./style.css";
 
-const SearchForm = props => {
-    return (
-        <form>
-            <div className="form-group">
-                <label className="VacationSearchLabel"></label>
-                <br></br>
-                <input className="col-12 form-control"
-                    value={props.search}
-                    type="text"
-                    name="searchVacation"
-                    placeholder="Enter Vacation Place"
-                    onChange={props.handleInputChange}
-                />
-            </div>
-            <button type="submit" className="submitBtn btn btn-primary" onClick={props.handleFormSubmit}>
-                Submit
-            </button>
-        </form>
-    )
-}
+// const SearchForm = props => {
+//     return (
+//         <form>
+//             <div className="form-group">
+//                 <label className="VacationSearchLabel"></label>
+//                 <br></br>
+//                 <input className="col-12 form-control"
+//                     value={props.search}
+//                     type="text"
+//                     name="searchVacation"
+//                     placeholder="Enter Vacation Place"
+//                     onChange={props.handleInputChange}
+//                 />
+//             </div>
+//             <button type="submit" className="submitBtn btn btn-primary" onClick={props.handleFormSubmit}>
+//                 Submit
+//             </button>
+//         </form>
+//     )
+// }
 
 
 
-export default SearchForm;
+// export default SearchForm;
