+++ client/src/pages/SearchVacations.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import API from "../utils/API";
 import Jumbotron from "../components/Jumbotron";
 import { Container, Row, Col } from "../components/Grid";
-import SearchForm from "../components/SearchForm";
+// import SearchForm from "../components/SearchForm";
 import SearchResult from "../components/SearchResult"
 import './styles.css';
 import { List, ListItem } from "../components/List"
@@ -119,12 +119,19 @@ class searchVacations extends Component {
             <Container fluid>
      
 
-                <Jumbotron />
-    
+                <Jumbotron
+                
+                handleFormSubmit={this.handleFormSubmit}
+                handleInputChange={this.handleInputChange}
+                
+                
+                
+                />
+{/*     
                 <SearchForm
                                 handleFormSubmit={this.handleFormSubmit}
                                 handleInputChange={this.handleInputChange}
-                            />
+                            /> */}
